+++ start_test.go
@@ -1,3 +1,10 @@
+// Copyright (c) Christoph Berger. All rights reserved.
+// Use of this source code is governed by the BSD (3-Clause)
+// License that can be found in the LICENSE.txt file.
+//
+// This source code may import third-party source code whose
+// licenses are provided in the respective license files.
+
 package start
 
 import (
@@ -24,7 +31,7 @@ func TestParse(t *testing.T) {
 		intFlag := flag.IntP("anint", "i", 23, "An integer flag")
 		boolFlag := flag.BoolP("anewbool", "b", true, "A boolean flag")
 
-		UseConfigFile("test/test.toml")
+		SetConfigFile("test/test.toml")
 		os.Setenv("START_ASTRING", "From Environment Variable")
 		Parse()
 		Convey("Then Parse() should find the correct values from config file, env var, or default. (Restriction: passing the command line flags is not possible with automated calls to go test)", func() {
@@ -46,6 +53,11 @@ func TestUp(t *testing.T) {
 	var second int
 	var global int
 	var params []string
+
+	SetInitFunc(func() error {
+		SetDescription("Testing testcmd")
+		return nil
+	})
 
 	// ContinueOnError is required when running goconvey as server; otherwise, unrecognized
 	// flags that are passed to the test executable will cause an error:
@@ -75,14 +87,32 @@ func TestUp(t *testing.T) {
 		},
 	})
 
+	Up()
+
 	Convey("The test command should read all flags and parameters.", t, func() {
-		err := Up()
 		So(first, ShouldEqual, 10)
 		So(second, ShouldEqual, 20)
 		So(global, ShouldEqual, 3)
 		So(params[0], ShouldEqual, "arg1")
 		So(params[1], ShouldEqual, "arg2")
-		So(err, ShouldEqual, nil)
-
+		So(description, ShouldEqual, "Testing testcmd")
 	})
+
+	Convey("The test command should contain only its private flags.", t, func() {
+		cmd := Commands["testcmd"]
+		So(len(cmd.Flags), ShouldEqual, 2)
+		So(cmd.Flags[0], ShouldEqual, "first")
+		So(cmd.Flags[1], ShouldEqual, "second")
+	})
+
+	Convey("The help command should be available", t, func() {
+		cmd := Commands["help"]
+		So(cmd, ShouldNotBeNil)
+		So(cmd.Name, ShouldEqual, "help")
+		So(cmd.Cmd, ShouldEqual, help)
+	})
+
+
 }
+	
+