+++ CHANGES.md
@@ -2,8 +2,10 @@ Changelog
 =========
 
 
-0.1.0
+0.1
 -----
+
+### 0.1.0
 
 Parse() implemented (passing all tests).
 
@@ -11,7 +13,25 @@ Parse() implemented (passing all tests).
 
 Fix: Case & tests for empty file name
 
-0.2.0
+0.2
 -----
 
+### 0.2.0
+
 Added: Command & Subcommand handling, Up() (passing all tests.)
+
+### 0.2.1
+
+Fix: Visibility issues
+
+### 0.2.2
+
+Added: Global init function.
+
+### 0.3.0
+
+Added: help command.
+Added: Some Set* functions.
+Added: Support for custom global init function.
+Added: Example in examples/test
+Fixed: empty usage